pyproject.toml
[project]
name = "rodmap"
version = "0.1.0"
description = "Backend RAG con Azure AI Search, LangGraph y Structured Output"
requires-python = ">=3.12,<3.13"
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.13.2",
    "aiosignal==1.4.0",
    "annotated-types==0.7.0",
    "anyio==3.7.1",
    "attrs==25.4.0",
    "authlib>=1.6.6",
    "azure-common==1.1.28",
    "azure-core==1.37.0",
    "azure-identity==1.15.0",
    "azure-search-documents==11.6.0",
    "certifi==2025.11.12",
    "cffi==2.0.0",
    "charset-normalizer==3.4.4",
    "click==8.3.1",
    "colorama==0.4.6",
    "cryptography==46.0.3",
    "dataclasses-json==0.6.7",
    "distro==1.9.0",
    "fastapi>=0.115.0",
    "frozenlist==1.8.0",
    "google-auth>=2.47.0",
    "greenlet==3.3.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "httpx-sse==0.4.3",
    "idna==3.11",
    "isodate==0.7.2",
    "jiter==0.12.0",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "langchain==1.2.0",
    "langchain-classic==1.0.1",
    "langchain-community==0.4.1",
    "langchain-core==1.2.6",
    "langchain-openai==1.1.6",
    "langchain-text-splitters==1.1.0",
    "langgraph==1.0.5",
    "langgraph-checkpoint==3.0.1",
    "langgraph-checkpoint-sqlite>=3.0.2",
    "langgraph-cli[inmem]>=0.2.6",
    "langgraph-prebuilt==1.0.5",
    "langgraph-sdk==0.3.1",
    "langsmith==0.5.0",
    "marshmallow==3.26.1",
    "msal==1.34.0",
    "msal-extensions==1.3.1",
    "multidict==6.7.0",
    "mypy-extensions==1.1.0",
    "numpy==1.26.4",
    "openai==1.109.1",
    "orjson==3.11.5",
    "ormsgpack==1.12.1",
    "packaging==23.2",
    "propcache==0.4.1",
    "pycparser==2.23",
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "pydantic-settings==2.12.0",
    "pyjwt==2.10.1",
    "python-dotenv==1.0.0",
    "pyyaml==6.0.3",
    "regex==2025.11.3",
    "requests==2.32.5",
    "requests-toolbelt==1.0.0",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.45",
    "starlette>=0.38.6",
    "tenacity==8.5.0",
    "tiktoken==0.12.0",
    "tqdm==4.67.1",
    "typing-extensions==4.15.0",
    "typing-inspect==0.9.0",
    "typing-inspection==0.4.2",
    "urllib3==2.6.2",
    "uuid-utils==0.12.0",
    "uvicorn>=0.26.0",
    "xxhash==3.6.0",
    "yarl==1.22.0",
    "zstandard==0.25.0",
]
